requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

pypdf

qdrant-client

sqlalchemy
psycopg2-binary

sentence-transformers
torch
